+++ packages/core/src/einheiten.ts
@@ -10,7 +10,11 @@ export type EinheitId =
   | "abteilung"
   | "grossverband";
 
-export type Einheit = Renderable & {
+export type EinheitRenderProps = {
+  textColor?: string;
+};
+
+export type Einheit = Renderable<EinheitRenderProps> & {
   id: EinheitId;
   label: string;
 };
@@ -60,14 +64,29 @@ export const einheiten: Array<Einheit> = [
     id: "zugtrupp",
     label: "Zugtrupp",
     size: [16, 7],
-    render: (svg) =>
+    render: (svg, props) =>
       svg
         .g()
         .attr("fill", "black")
         .push(svg.circle([8, 3], 2))
-        .push(svg.circle([2, 12], 1.5))
-        .push(svg.circle([8, 12], 1.5))
-        .push(svg.circle([14, 12], 1.5)),
+        .push(
+          svg
+            .circle([2, 12], 1.5)
+            .attr("stroke", props?.textColor)
+            .attr("fill", props?.textColor)
+        )
+        .push(
+          svg
+            .circle([8, 12], 1.5)
+            .attr("stroke", props?.textColor)
+            .attr("fill", props?.textColor)
+        )
+        .push(
+          svg
+            .circle([14, 12], 1.5)
+            .attr("stroke", props?.textColor)
+            .attr("fill", props?.textColor)
+        ),
   },
   {
     id: "bereitschaft",